gestita directory temporanea per dump db

+++ db-backup/mysqlbackup.py
@@ -9,6 +9,7 @@ from mysqlbackupcfg import config
 
 # get configuration parameters
 db_to_backup = config['db_to_backup']
+backup_direct = config['backup_direct']
 tmp_export_dir = config['tmp_export_dir']
 backup_dir = config['backup_dir']
 backup_dir_copy = config['backup_dir_copy']
@@ -27,16 +28,25 @@ mail_to = config['mail_to']
 
 # prepare result message
 detail_messages = []
-esito = ''
-if (os.path.exists(tmp_export_dir) and os.path.isdir(tmp_export_dir)):
-    # temporary export dir exists
-    # remove all file from temporary export dir
-    for item in os.listdir(tmp_export_dir):
-        # check dir content
-        item_path = tmp_export_dir + '/' + item
-        if (os.path.isfile(item_path)):
-            # is a file, remove it
-            os.remove(item_path)
+error = False
+export_dir = backup_dir
+if not backup_direct: 
+    if (os.path.exists(tmp_export_dir) and os.path.isdir(tmp_export_dir)):
+        # temporary export dir exists
+        export_dir = tmp_export_dir 
+        # remove all file from temporary export dir
+        for item in os.listdir(tmp_export_dir):
+            # check dir content
+            item_path = tmp_export_dir + '/' + item
+            if (os.path.isfile(item_path)):
+                # is a file, remove it
+                os.remove(item_path)
+    else:
+        # temprary export dir not fouded
+        error = True
+        detail_messages.append('La directory temporanea di export ' + backup_dir + ' non esiste.')
+
+if not error:
     if (os.path.exists(backup_dir) and os.path.isdir(backup_dir)):
         # backup dir exists
         # remove old file from backup dir
@@ -77,50 +87,50 @@ if (os.path.exists(tmp_export_dir) and os.path.isdir(tmp_export_dir)):
                 log_name = start_time.strftime('%Y%m%d') + '_daily-' + filename_prefix + '_' + server + '_' + db_name + '.log'
                 # dump db
                 result = os.system('mysqldump -u ' + user + ' -p' + password + ' -h ' + server + ' -P ' + port + ' ' +
-                                   db_name + ' 2> ' + tmp_export_dir + '/' + log_name + ' | gzip > ' + tmp_export_dir + '/' + dump_name)
+                                    db_name + ' 2> ' + export_dir + '/' + log_name + ' | gzip > ' + export_dir + '/' + dump_name)
                 # prepare result message
                 message = 'Backup del db ' + db_name + ' dal server ' + server + '.'
                 if result != 0:
                     # error
-                    esito = 'ERRORE!!!'
+                    error = True
                     detail_messages.append(message + ' ERRORE!!!!')
                 else:
                     # success
                     # check log file
-                    log_size = os.path.getsize(tmp_export_dir + '/' + log_name)
+                    log_size = os.path.getsize(export_dir + '/' + log_name)
                     if log_size > 0:
                         # error
-                        esito = 'ERRORE!!!'
+                        error = True
                         detail_messages.append(message + ' ERRORE!!!! Controllare file di log.')
                     else:
                         # success
-                        esito = 'OK'
+                        error = False
                         detail_messages.append(message + ' OK!')
                         # copy dump file to destination dir
                         if (backup_dir_copy == True):
                             # copy dump file to destination dir
-                            sys_command = backup_dir_copy_cmd.replace('[dump_file]', tmp_export_dir + '/' + dump_name)
+                            sys_command = backup_dir_copy_cmd.replace('[dump_file]', export_dir + '/' + dump_name)
                             result = os.system(sys_command)
                             message = 'Copia del dump file ' + db_name + '.'
                             if result != 0:
                                 # error
-                                esito = 'ERRORE!!!'
+                                error = True
                                 detail_messages.append(message + ' ERRORE!!!!')
                             else:
                                 # success
-                                esito = 'OK'
+                                error = False
                                 detail_messages.append(message + ' OK!')
                             # copy log file to destination dir
-                            sys_command = backup_dir_copy_cmd.replace('[dump_file]', tmp_export_dir + '/' + log_name)
+                            sys_command = backup_dir_copy_cmd.replace('[dump_file]', export_dir + '/' + log_name)
                             result = os.system(sys_command)
                             message = 'Copia del log file ' + db_name + '.'
                             if result != 0:
                                 # error
-                                esito = 'ERRORE!!!'
+                                error = True
                                 detail_messages.append(message + ' ERRORE!!!!')
                             else:
                                 # success
-                                esito = 'OK'
+                                error = False
                                 detail_messages.append(message + ' OK!')
                             if (weekly_backup == True and current_date.weekday() == 6):
                                 # rename to weekly file
@@ -128,16 +138,16 @@ if (os.path.exists(tmp_export_dir) and os.path.isdir(tmp_export_dir)):
                                 os.rename(dump_name, new_name)
                                 dump_name = new_name
                                 # copy file to destination dir
-                                sys_command = backup_dir_copy_cmd.replace('[dump_file]', tmp_export_dir + '/' + dump_name)
+                                sys_command = backup_dir_copy_cmd.replace('[dump_file]', export_dir + '/' + dump_name)
                                 result = os.system(sys_command)
                                 message = 'Copia del dump file settimanale ' + db_name + '.'
                                 if result != 0:
                                     # error
-                                    esito = 'ERRORE!!!'
+                                    error = True
                                     detail_messages.append(message + ' ERRORE!!!!')
                                 else:
                                     # success
-                                    esito = 'OK'
+                                    error = False
                                     detail_messages.append(message + ' OK!')
                             if (monthly_backup == True and current_date.day == 1):
                                 # rename to monthly file
@@ -145,16 +155,16 @@ if (os.path.exists(tmp_export_dir) and os.path.isdir(tmp_export_dir)):
                                 os.rename(dump_name, new_name)
                                 dump_name = new_name
                                 # copy file to destination dir
-                                sys_command = backup_dir_copy_cmd.replace('[dump_file]', tmp_export_dir + '/' + dump_name)
+                                sys_command = backup_dir_copy_cmd.replace('[dump_file]', export_dir + '/' + dump_name)
                                 result = os.system(sys_command)
                                 message = 'Copia del dump file mensile ' + db_name + '.'
                                 if result != 0:
                                     # error
-                                    esito = 'ERRORE!!!'
+                                    error = True
                                     detail_messages.append(message + ' ERRORE!!!!')
                                 else:
                                     # success
-                                    esito = 'OK'
+                                    error = False
                                     detail_messages.append(message + ' OK!')
                             if (yearly_backup == True and current_date.day == 1 and current_date.month == 1):
                                 # rename to yearly file
@@ -162,28 +172,29 @@ if (os.path.exists(tmp_export_dir) and os.path.isdir(tmp_export_dir)):
                                 os.rename(dump_name, new_name)
                                 dump_name = new_name
                                 # copy file to destination dir
-                                sys_command = backup_dir_copy_cmd.replace('[dump_file]', tmp_export_dir + '/' + dump_name)
+                                sys_command = backup_dir_copy_cmd.replace('[dump_file]', export_dir + '/' + dump_name)
                                 result = os.system(sys_command)
                                 message = 'Copia del dump file annuale ' + db_name + '.'
                                 if result != 0:
                                     # error
-                                    esito = 'ERRORE!!!'
+                                    error = True
                                     detail_messages.append(message + ' ERRORE!!!!')
                                 else:
                                     # success
-                                    esito = 'OK'
+                                    error = False
                                     detail_messages.append(message + ' OK!')
+                        if not backup_direct: 
+                            if (os.path.exists(tmp_export_dir) and os.path.isdir(tmp_export_dir)):
+                                # remove files from temporary export dir
+                                os.remove(export_dir + '/' + dump_name)
+                                os.remove(export_dir + '/' + log_name)
                 end_time = datetime.now()
                 detail_messages.append('Inizio: ' + start_time.strftime('%d-%m-%Y %H:%M') + ' - Fine: ' + end_time.strftime('%d-%m-%Y %H:%M') +
-                                       ' - Dimensione: ' + str(os.path.getsize(tmp_export_dir + '/' + dump_name)) + '\n')
+                                        ' - Dimensione: ' + str(os.path.getsize(export_dir + '/' + dump_name)) + '\n')                
     else:
         # backup dir not fouded
-        esito = 'ERRORE!!!'
+        error = True    
         detail_messages.append('La directory di backup ' + backup_dir + ' non esiste.')
-else:
-    # temprary export dir not fouded
-    esito = 'ERRORE!!!'
-    detail_messages.append('La directory temporanea di export ' + backup_dir + ' non esiste.')
 
 # send email ####
 if len(detail_messages) > 0:
@@ -195,7 +206,7 @@ if len(detail_messages) > 0:
         server.login(mail_server_user, mail_server_password)
         # prepare mail message
         message = []
-        message.append('Subject: [' + socket.gethostname() + '] Backup dei database - Esito: ' + esito)
+        message.append('Subject: [' + socket.gethostname() + '] Backup dei database - Esito: ' + ('ERROR!!!' if error else 'OK'))
         message.append('From: ' + mail_from)
         message.append('To: ' + ', '.join(mail_to))
         message.append('Backup dei database del server ' + socket.gethostname() + '\n')
